requirements.txt
numpy>=1.23
scipy>=1.10
matplotlib>=3.7
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
torch>=2.0
